+++ src/components/global/Header/index.tsx
@@ -1,5 +1,4 @@
 import { ColorTheme } from "./ColorTheme";
 import { Lang } from "./Lang";
-import { Router } from "./Router";
 
-export { ColorTheme, Lang, Router };
+export { ColorTheme, Lang };
